Pre-commit run - ICON4Py Quality Assurance

+++ model/atmosphere/advection/src/icon4py/model/atmosphere/advection/upwind_hflux_miura_cycl_stencil_03b.py
@@ -24,9 +24,9 @@ def _upwind_hflux_miura_cycl_stencil_03b(
     z_tracer_mflx_2_dsl: Field[[EdgeDim, KDim], float],
     z_tracer_mflx_3_dsl: Field[[EdgeDim, KDim], float],
 ) -> Field[[EdgeDim, KDim], float]:
-    p_out_e = (z_tracer_mflx_1_dsl
-             + z_tracer_mflx_2_dsl 
-             + z_tracer_mflx_3_dsl)/float(3)
+    p_out_e = (z_tracer_mflx_1_dsl + z_tracer_mflx_2_dsl + z_tracer_mflx_3_dsl) / float(
+        3
+    )
     return p_out_e
 
 
@@ -38,8 +38,5 @@ def upwind_hflux_miura_cycl_stencil_03b(
     p_out_e: Field[[EdgeDim, KDim], float],
 ):
     _upwind_hflux_miura_cycl_stencil_03b(
-        z_tracer_mflx_1_dsl,
-        z_tracer_mflx_2_dsl,
-        z_tracer_mflx_3_dsl,
-        out=(p_out_e)
+        z_tracer_mflx_1_dsl, z_tracer_mflx_2_dsl, z_tracer_mflx_3_dsl, out=(p_out_e)
     )
